requirements.txt
azure-core==1.36.0
azure-identity==1.25.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
diskcache==5.6.3
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
json_repair==0.53.0
lazy_loader==0.4
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
networkx==3.5
ninja==1.13.0
numpy==2.2.6
Office365-REST-Python-Client==2.6.2
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pyclipper==1.3.0.post6
pycparser==2.23
PyJWT==2.10.1
PyMuPDF==1.26.6
python-bidi==0.6.7
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
